made a component folder and moved ColorBlock and ColorForms to components folder
